Velho_a_Novo

+++ Old/src/Manager.Services/DTO/UserDTO.cs
@@ -2,7 +2,7 @@ using System.Text.Json.Serialization;
 
 namespace Manager.Services.DTO{
     public class UserDTO{
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Email { get; set; }
 
@@ -12,7 +12,7 @@ namespace Manager.Services.DTO{
         public UserDTO()
         {}
 
-        public UserDTO(long id, string name, string email, string password)
+        public UserDTO(int id, string name, string email, string password)
         {
             Id = id;
             Name = name;
